Fixing up unique id generator for new todo items

+++ Ianf.Global360ToDo.Domain/ToDo.cs
@@ -3,7 +3,7 @@
 public readonly record struct ToDo
 {
     public ToDo() { }
-    public long Id { get; init; } = 0;
+    public int Id { get; init; } = 0;
     public string Title { get; init; } = string.Empty;
     public string Contents { get; init; } = string.Empty;
 }
