+++ Ianf.Global360ToDo.Repositories/IToDoRepository.cs
@@ -4,7 +4,7 @@ namespace Ianf.Global360ToDo.Repositories;
 
 public interface IToDoRepository
 {
-    public Task<long> AddToDoItem(NewToDo toDo);
-    public Task<List<ToDo>> RemoveToDoItem(long toDoId);
+    public Task<int> AddToDoItem(NewToDo toDo);
+    public Task<List<ToDo>> RemoveToDoItem(int toDoId);
     public Task<List<ToDo>> GetToDoItems();
 }
